--- app/services/offers/offer_total_updater_service.rb
@@ -1,14 +1,0 @@
-module Offers
-  class OfferTotalUpdaterService < BaseTotalCalculatorService
-    def initialize(offer)
-      @pending_offer = offer
-      @order = offer.order
-    end
-
-    def process!
-      set_offer_totals!
-    end
-
-    attr_reader :offer
-  end
-end
